requirements.txt
pymongo>=4.6.0
requests>=2.31.0
beautifulsoup4>=4.12.0
fastapi>=0.104.0
uvicorn>=0.24.0
schedule>=1.2.0
python-dotenv>=1.0.0
